+++ docs/src/examples/components/Divider/Variations/DividerExampleSize.tsx
@@ -3,8 +3,8 @@ import React from 'react'
 import { Divider } from 'stardust'
 
 const DividerExampleSize = () =>
-  _.times(10, i => {
-    const size = i + 1
+  _.times(11, i => {
+    const size = i
 
     return (
       <Divider key={size} size={size}>